fix(provider): `marker` provider window
Bug: sometimes the `marker` provider tries to query window markers whose
ID is '-1'. This window does not exist, so trying to access it results in
some errors

Cause: The function `require('ufo.utils').getWinByBuf()` returns '-1'
when it cannot return a window ID for the given buffer. The old
implementation of the `Marker.getFolds()` function did not check if the
window exists and tries to access the window whose ID is '-1'

Solution: Check that the `require('ufo.utils').getWinByBuf()` function
returns a valid window ID (positive number). If it is not valid, the
function does not query the buffer folds and returns nothing