feat(docs): mark UfoConfig fields as optional
Mark fields `open_fold_hl_timeout`, `close_fold_kinds_for_ft`,
`enable_get_fold_virt_text`, and `preview` in the `UfoConfig` as
optional.

This allows the user to annotate his config options as:
```lua
---@type UfoConfig
opts = {
    -- ...
}
```
and rely on the defined default values in the plugin. This is beneficial
for the user because it allows completion of the fields of the config
without giving diagnostic warnings.