fix(fold): fold text after vim highlighter parsing state change (#305)
Problem: The fold text was broken (shows 0 instead of the buffer text)
after `vim.treesitter.highlighter` changed its `parsing` property to be
a `table` instead of a `boolean` [0].

Solution: Adapt the code to detect when `parsing` is a table, and get
the necessary information from that table.

For backward compatibility, keep the old behavior if `parsing` it is not
a `table`.

[0]: https://github.com/neovim/neovim/commit/faae1e72a27b31017be639f24e3a9b374b4be255